Added class FamilyBuilder
